requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-modeltranslation==0.18.4
django-widget-tweaks==1.4.12
Pillow==9.2.0
python-decouple==3.6
sqlparse==0.4.2
